+++ libft/ft_strlcat.c
@@ -3,20 +3,20 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strlcat.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jlimones <jlimones@student.42.fr>          +#+  +:+       +#+        */
+/*   By: user <user@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/22 22:38:32 by jlimones          #+#    #+#             */
-/*   Updated: 2022/09/23 10:13:07 by jlimones         ###   ########.fr       */
+/*   Updated: 2023/05/21 04:30:12 by user             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t	ft_strlcat(char *dst, char *src, size_t size)
+int	ft_strlcat(char *dst, char *src, int size)
 {
-	size_t	i;
-	size_t	y;
-	size_t	rt;
+	int	i;
+	int	y;
+	int	rt;
 
 	i = 0;
 	y = ft_strlen(dst);
